requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.12.1

# Configuration and validation
pydantic==2.4.2
pydantic-settings==2.0.3

# Logging
structlog==23.2.0

# HTTP client
httpx==0.25.2

# Monitoring and metrics
prometheus-client==0.19.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Code quality
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Git operations (optional, for local git integration)
gitpython==3.1.40

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
